detail page view dialog added
