+++ lib/features/v2/details/view/event_detail_view.dart
@@ -7,7 +7,6 @@ import 'package:life_shared/life_shared.dart';
 import 'package:vbaseproject/features/v2/details/mixin/event_detail_mixin.dart';
 import 'package:vbaseproject/product/init/language/locale_keys.g.dart';
 import 'package:vbaseproject/product/model/enum/text_field/text_field_max_lengths.dart';
-import 'package:vbaseproject/product/package/image/custom_network_image.dart';
 import 'package:vbaseproject/product/utility/constants/app_constants.dart';
 import 'package:vbaseproject/product/utility/constants/app_icons.dart';
 import 'package:vbaseproject/product/utility/extension/index.dart';
@@ -17,6 +16,7 @@ import 'package:vbaseproject/product/widget/dialog/general_text_dialog.dart';
 import 'package:vbaseproject/product/widget/dialog/sub_widget/general_dialog_button.dart';
 import 'package:vbaseproject/product/widget/general/index.dart';
 import 'package:vbaseproject/product/widget/general/title/general_content_small_title.dart';
+import 'package:vbaseproject/product/widget/image/custom_image_with_view_dialog.dart';
 import 'package:vbaseproject/product/widget/size/widget_size.dart';
 import 'package:vbaseproject/product/widget/text/title_description_text.dart';
