+++ lib/features/v2/details/view/widget/news_detail_sub_view.dart
@@ -15,34 +15,13 @@ final class _ImageWithButtonAndNameStack extends StatelessWidget {
     return Stack(
       clipBehavior: Clip.none,
       children: [
-        _ImageSizedBox(image: image),
+        CustomImageWithViewDialog(image: image),
         SafeArea(
           child: _BackButtonContainer(
             onPressed: backButtonAction,
           ),
         ),
       ],
-    );
-  }
-}
-
-@immutable
-final class _ImageSizedBox extends StatelessWidget {
-  const _ImageSizedBox({
-    required this.image,
-  });
-
-  final String image;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: context.sized.width,
-      height: context.sized.dynamicHeight(0.3),
-      child: CustomNetworkImage(
-        imageUrl: image,
-        fit: BoxFit.cover,
-      ),
     );
   }
 }
