+++ lib/product/widget/dialog/index.dart
@@ -5,7 +5,7 @@ export 'general_icon_text_dialog.dart';
 export 'general_text_dialog.dart';
 export 'not_connected_to_internet_dialog.dart';
 export 'pdf_preview_dialog.dart';
-export 'phone_zoom_dialog.dart';
+export 'photo_view_dialog.dart';
 export 'success_data_posted_dialog.dart';
 export 'success_scholarship_posted_dialog.dart';
 export 'video_dialog.dart';
