+++ lib/product/widget/image/custom_image_with_view_dialog.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import 'package:kartal/kartal.dart';
+import 'package:vbaseproject/product/package/image/custom_network_image.dart';
+import 'package:vbaseproject/product/widget/dialog/photo_view_dialog.dart';
+
+@immutable
+final class CustomImageWithViewDialog extends StatelessWidget {
+  const CustomImageWithViewDialog({
+    required this.image,
+    super.key,
+  });
+
+  final String image;
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: () {
+        PhotoViewDialog(imageUrl: image).show(context);
+      },
+      child: SizedBox(
+        width: context.sized.width,
+        height: context.sized.dynamicHeight(0.3),
+        child: CustomNetworkImage(
+          imageUrl: image,
+          fit: BoxFit.cover,
+        ),
+      ),
+    );
+  }
+}
